[#13] Minor source code improvements (Part 45: Resolution of compiler warnings C4244)

+++ PropGrid/src/PropGrid.cpp
@@ -110,7 +110,7 @@
 
 
 
-	TPropGridInfo::TPropGridInfo(HWND hPropGrid,WORD propGridWidth,WORD propGridHeight)
+	TPropGridInfo::TPropGridInfo(HWND hPropGrid,LONG propGridWidth,LONG propGridHeight)
 		// ctor
 		: handle(hPropGrid)
 		, listBox(this,propGridWidth,propGridHeight)
